Change the structure of bloc_presentation_test example
